conditional delete and some proptypes

+++ Bloglist/front-end/src/components/Blog.js
@@ -6,12 +6,19 @@ class Blog extends React.Component {
     super(props)
     this.state = {
       disp: ['', 'none'],
-      value: 1
+      blogVisibility: 1,
+      buttonVisibility: 1
     }
   }
 
   toggle = () => {
-    this.setState({value: 1 - this.state.value})
+    const username = this.props.blog.user.username
+    const logged = this.props.username
+
+    if(username === logged) {
+      this.setState({buttonVisibility: 1 - this.state.buttonVisibility})
+    }
+    this.setState({blogVisibility: 1 - this.state.blogVisibility})
   }
 
   async onLike(blog) {
@@ -27,9 +34,14 @@ class Blog extends React.Component {
 
   async onDelete(blog) {
     try {
-      await blogService
-        .remove(blog.id)
-      this.props.onDelete(blog.id)
+      const res = window.confirm(
+        `Do you want to delete '${blog.title}' by ${blog.author}?`
+      )
+      if(res) {
+        await blogService
+          .remove(blog.id)
+        this.props.onDelete(blog.id)
+      }
     } catch(e) {
       console.log(e)
     }
@@ -40,14 +52,16 @@ class Blog extends React.Component {
     const user = {...blog.user}
     return (
       <div className="blog">
-        <div onClick={this.toggle}>
+        <div key={blog.id} onClick={this.toggle}>
           {blog.title} by {blog.author}
         </div>
-        <div className="blogextra" style={{display: this.state.disp[this.state.value]}}>
+        <div className="blogextra" style={{display: this.state.disp[this.state.blogVisibility]}}>
           {blog.url}<br />
-          {blog.likes} likes<button onClick={() => this.onLike(blog)}>like</button><br />
-          Added by {user.name} <br />
-          <button onClick={() => this.onDelete(blog)}>Delete</button>
+          This blog has {blog.likes} likes<button onClick={() => this.onLike(blog)}>like</button><br />
+          Added by {user.name}
+          <div style={{display: this.state.disp[this.state.buttonVisibility]}}>
+            <button onClick={() => this.onDelete(blog)}>Delete</button>
+          </div>
         </div>
       </div>
     )
